[ios][tests] Re-land expo-updates migration with serialized integrity check
Reverts commit 48202c5f2ad and fixes the one suite that actually crashed instead of
keeping the whole package on Quick/Nimble.

Only `DatabaseIntegrityCheckTests` killed a worker: it built a fixed directory
(`applicationSupportDir/UpdatesDatabaseTests`) in `init` and force-removed it in
`deinit`, so parallel tests deleted it from under each other and the `try!` trapped.

`.serialized` alone isn't enough, since it only orders tests within a suite, and three
top-level suites shared that directory name. `UpdatesDatabaseTests` and
`FileDownloaderTests` only survive because they use `try?` throughout, which swallows
the race. So this also gives the suite its own directory name, following how
`UpdatesDatabaseTests` names one per nested suite, and relaxes the `deinit` removal to
`try?` to match `FileDownloaderTests`.

The `UpdatesStateMachine.swift` "invalid transition requested" assertion failures in the
CI logs were expected output from the retained Quick tests that rely on Nimble's
`throwAssertion()`; that worker exited cleanly. The other migrated files need no
serialization: the selection policy suites are pure value logic, and
`FileDownloaderManifestParsingTests` computes a directory path without creating or
deleting it.
