+++ packages/expo-updates/ios/Tests/UpdatesStateMachineSpec.swift
@@ -1,5 +1,10 @@
 //  Copyright (c) 2023 650 Industries, Inc. All rights reserved.
 
+// NOTE: The rest of the UpdatesStateMachine tests have been migrated to Swift Testing in
+// UpdatesStateMachineTests.swift. The two tests below remain on Quick/Nimble because they rely on
+// Nimble's `throwAssertion()` matcher (to catch Swift `assert`/`precondition`/`fatalError`), which
+// has no Swift Testing equivalent. Migrate these once such support exists.
+
 // swiftlint:disable closure_body_length
 
 import ExpoModulesTestCore
@@ -20,158 +25,6 @@ class TestStateChangeEventManager: UpdatesEventManager {
 class UpdatesStateMachineSpec: ExpoSpec {
   override class func spec() {
     describe("default state") {
-      it("instantiates") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-        expect(machine.getStateForTesting()) == .idle
-      }
-
-      it("sequence numbers") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-        expect(machine.getStateForTesting()) == .idle
-
-        expect(machine.context.sequenceNumber) == 0
-
-        machine.processEventForTesting(.startStartup)
-        machine.processEventForTesting(.check)
-        machine.processEventForTesting(.checkCompleteUnavailable)
-        machine.processEventForTesting(.endStartup)
-
-        expect(machine.context.sequenceNumber) == 4
-      }
-
-      it("restart") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-        expect(machine.getStateForTesting()) == .idle
-
-        expect(machine.context.isRestarting) == false
-        machine.processEventForTesting(.restart)
-        expect(machine.context.isRestarting) == true
-        expect(machine.context.sequenceNumber) == 1
-
-        machine.resetAndIncrementRestartCountForTesting()
-        expect(machine.context.restartCount) == 1
-        expect(machine.context.isRestarting) == false
-        expect(machine.context.sequenceNumber) == 2
-      }
-
-      it("should handle startStartup and endStartup") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-
-        machine.processEventForTesting(.startStartup)
-        expect(machine.getStateForTesting()) == .idle
-        expect(testStateChangeEventManager.lastContext?.isStartupProcedureRunning) == true
-
-        machine.processEventForTesting(.endStartup)
-        expect(machine.getStateForTesting()) == .idle
-        expect(testStateChangeEventManager.lastContext?.isStartupProcedureRunning) == false
-      }
-
-      it("should handle check and checkCompleteAvailable") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-
-        machine.processEventForTesting(.check)
-        expect(machine.getStateForTesting()) == .checking
-        expect(testStateChangeEventManager.lastContext?.isChecking) == true
-
-        machine.processEventForTesting(.checkCompleteWithUpdate(manifest: ["updateId": "0000-xxxx"]))
-        expect(machine.getStateForTesting()) == .idle
-        expect(machine.context.isChecking) == false
-        expect(machine.context.checkError).to(beNil())
-        expect(machine.context.latestManifest?["updateId"] as? String ?? "") == "0000-xxxx"
-        expect(machine.context.isUpdateAvailable) == true
-        expect(machine.context.isUpdatePending) == false
-        expect(testStateChangeEventManager.lastContext?.isUpdateAvailable) == true
-        let values = testStateChangeEventManager.lastContext
-        expect(values?.isUpdateAvailable) == true
-      }
-
-      it("should handle check and checkCompleteUnavailable") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-
-        machine.processEventForTesting(.check)
-        expect(machine.getStateForTesting()) == .checking
-
-        machine.processEventForTesting(.checkCompleteUnavailable)
-        expect(machine.getStateForTesting()) == .idle
-        expect(machine.context.isChecking) == false
-        expect(machine.context.checkError).to(beNil())
-        expect(machine.context.latestManifest).to(beNil())
-        expect(machine.context.isUpdateAvailable) == false
-        expect(machine.context.isUpdatePending) == false
-      }
-
-      it("should handle a completed download with an update") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-
-        machine.processEventForTesting(.download)
-        expect(machine.getStateForTesting()) == .downloading
-
-        machine.processEventForTesting(.downloadCompleteWithUpdate(manifest: ["updateId": "0000-xxxx"]))
-        expect(machine.getStateForTesting()) == .idle
-        expect(machine.context.isChecking) == false
-        expect(machine.context.downloadError).to(beNil())
-        expect(machine.context.latestManifest?["updateId"] as? String ?? "") == "0000-xxxx"
-        expect(machine.context.downloadedManifest?["updateId"] as? String ?? "") == "0000-xxxx"
-        expect(machine.context.isUpdateAvailable) == true
-        expect(machine.context.isUpdatePending) == true
-        expect(machine.context.rollback) == nil
-      }
-
-      it("should handle a completed download with a rollback") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-
-        machine.processEventForTesting(.download)
-        expect(machine.getStateForTesting()) == .downloading
-
-        machine.processEventForTesting(.downloadCompleteWithRollback)
-        expect(machine.getStateForTesting()) == .idle
-        expect(machine.context.isDownloading) == false
-        expect(machine.context.downloadError).to(beNil())
-        expect(machine.context.isUpdatePending) == true
-      }
-
-      it("should handle download progress") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-
-        machine.processEventForTesting(.download)
-        expect(machine.getStateForTesting()) == .downloading
-        expect(testStateChangeEventManager.lastContext?.downloadProgress) == 0
-
-        machine.processEventForTesting(.downloadProgress(progress: 0.5))
-        expect(machine.getStateForTesting()) == .downloading
-        expect(testStateChangeEventManager.lastContext?.downloadProgress) == 0.5
-
-        machine.processEventForTesting(.downloadCompleteUnavailable)
-        expect(machine.getStateForTesting()) == .idle
-        expect(testStateChangeEventManager.lastContext?.downloadProgress) == 1
-      }
-
-      it("should handle rollback") {
-        let testStateChangeEventManager = TestStateChangeEventManager()
-        let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
-        let commitTime = Date()
-        machine.processEventForTesting(.check)
-        expect(machine.getStateForTesting()) == .checking
-
-        machine.processEventForTesting(.checkCompleteWithRollback(rollbackCommitTime: commitTime))
-        expect(machine.getStateForTesting()) == .idle
-        expect(machine.context.isChecking) == false
-        expect(machine.context.checkError).to(beNil())
-        expect(machine.context.latestManifest).to(beNil())
-        expect(machine.context.isUpdateAvailable) == true
-        expect(machine.context.isUpdatePending) == false
-        expect(machine.context.rollback?.commitTime) == commitTime
-      }
-
       it("invalid transitions are handled as expected") {
         let testStateChangeEventManager = TestStateChangeEventManager()
         let machine = UpdatesStateMachine(logger: UpdatesLogger(), eventManager: testStateChangeEventManager, validUpdatesStateValues: Set(UpdatesStateValue.allCases))
